chitty/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.7
redis==5.0.1
minio==7.1.17
cryptography==41.0.4
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3